update
changed all Double to Float

+++ app/src/main/java/com/example/foodinfo/local/entity/BaseFieldFilterEntity.kt
@@ -22,10 +22,10 @@ data class BaseFieldFilterEntity(
     val filterName: String,
 
     @ColumnInfo(name = Columns.MIN_VALUE)
-    val minValue: Double,
+    val minValue: Float,
 
     @ColumnInfo(name = Columns.MAX_VALUE)
-    val maxValue: Double
+    val maxValue: Float
 
 ) {
     object Columns {
