+++ app/src/main/java/com/example/foodinfo/repository/model/filter_field/NutrientFilterField.kt
@@ -3,6 +3,6 @@ package com.example.foodinfo.repository.model.filter_field
 
 data class NutrientFilterField(
     val name: String,
-    val minValue: Double? = null,
-    val maxValue: Double? = null
+    val minValue: Float? = null,
+    val maxValue: Float? = null
 )
